+++ lib/eth.rb
@@ -35,11 +35,11 @@ require "eth/contract/encoder"
 require "eth/contract/decoder"
 require "eth/contract/formatter"
 require "eth/contract/deployment"
+require "eth/contract/initializer"
 require "eth/client"
 require "eth/client/http"
 require "eth/client/ipc"
 require "eth/eip712"
-require "eth/initializer"
 require "eth/key"
 require "eth/rlp"
 require "eth/signature"
